+++ hooks/lib/managers/menumanager.lua
@@ -515,31 +515,31 @@ Hooks:Add("MenuManagerInitialize", "MenuManagerInitialize_UltimateTrainer", func
     end
 
     MenuCallbackHandler.ut_time_set_environment_early_morning = function(self, item)
-        UT.Time:setEnvironment("environments/pd2_env_hox_02/pd2_env_hox_02")
+        UT.Time:setEnvironmentSetting("environments/pd2_env_hox_02/pd2_env_hox_02")
     end
 
     MenuCallbackHandler.ut_time_set_environment_morning = function(self, item)
-        UT.Time:setEnvironment("environments/pd2_env_morning_02/pd2_env_morning_02")
+        UT.Time:setEnvironmentSetting("environments/pd2_env_morning_02/pd2_env_morning_02")
     end
 
     MenuCallbackHandler.ut_time_set_environment_mid_day = function(self, item)
-        UT.Time:setEnvironment("environments/pd2_env_mid_day/pd2_env_mid_day")
+        UT.Time:setEnvironmentSetting("environments/pd2_env_mid_day/pd2_env_mid_day")
     end
 
     MenuCallbackHandler.ut_time_set_environment_afternoon = function(self, item)
-        UT.Time:setEnvironment("environments/pd2_env_afternoon/pd2_env_afternoon")
+        UT.Time:setEnvironmentSetting("environments/pd2_env_afternoon/pd2_env_afternoon")
     end
 
     MenuCallbackHandler.ut_time_set_environment_night = function(self, item)
-        UT.Time:setEnvironment("environments/pd2_env_n2/pd2_env_n2")
+        UT.Time:setEnvironmentSetting("environments/pd2_env_n2/pd2_env_n2")
     end
 
     MenuCallbackHandler.ut_time_set_environment_misty_night = function(self, item)
-        UT.Time:setEnvironment("environments/pd2_env_arm_hcm_02/pd2_env_arm_hcm_02")
+        UT.Time:setEnvironmentSetting("environments/pd2_env_arm_hcm_02/pd2_env_arm_hcm_02")
     end
 
     MenuCallbackHandler.ut_time_set_environment_foggy_night = function(self, item)
-        UT.Time:setEnvironment("environments/pd2_env_foggy_bright/pd2_env_foggy_bright")
+        UT.Time:setEnvironmentSetting("environments/pd2_env_foggy_bright/pd2_env_foggy_bright")
     end
 
     MenuCallbackHandler.ut_time_reset_environment = function(self, item)